Added direct manipulation of closure environments
Added support to the compiler for direct manipulation of
a closures environment.  This is supported through a dot syntax.
You can read and write to the environment using this dot syntax.

Reading: (closure.slot:type) Writing: (closure.slot:type value)

For example:

(definec test
  (lambda (a:i32)
    (let ((f (lambda ()
	       (* a a))))
      (printf "1: %d\n" (f))
      (f.a:i32 10)
      (printf "2: %d\n" (f)))))

(test 5)

Prints to the log

1: 25
2: 100

25 being f called with a's starting value of 5
100 being f called with a's value changed to 10

Ideally we would use type inferencing to infer the slots type
without the need to explicitly include it.  Additionally this
can be type checked at runtime although I have not added this yet.