Add scripts for incremental builds.
I have created a build directory for holding the object files and
a config directory for configuration files.

The config/llvm.bash can be edited to specify the LLVM install
location.  The config/{linux,darwin}.make files are used to
configure the build for the detected OS.

In addition to allowing incremental builds, the make scripts are
written to arrange automatic tracking of C++ include-dependencies.
For this task, I've added the -MMD option to CXXFLAGS and I have
included a small perl script to collect dependency rules into one
file.

The installation instructions in INSTALL have been updated.