Replace compiler assoc lists with native std::map
The primary change here is to replaces impc:ir:gname assoc lists (i.e. prev) with a native std::map. This is addmittedly a little ugly but is a short term hack designed to fit into the existing compiler with as little modification as possible.  This makes makes a substantial performance difference in the short term but will be replaced in the longer term.

This commit also removes many string-appends, no longer required since emit change.

There have been a lot of regex search and replaces in this change - everything appears to be working ok - fingers crossed.