Compiler Zone Changes
Changed some of the zoning behaviour, particularly related to the set!  Now set! modifies a closures environment variable in place.  This is obviously far more efficient and keeps alterations in the correct zone.  Unfortunately there is problem applying this to closure sets at preset.  Until I come up with a plan for fixing this the compiler will just leak closure sets.