Removed default audio input on linux
I changed my mind on this because of poor portaudio input performance.  This is probably because of pulse audio but for the moment I don't have time to investigate further so I'm removing this as the default.  It is simple enough for individuals to put back in if they want audio input on linux.