Added native array type to extempore
Extempore lang now supports first class arrays.
(make-array num type) now returns a pointer to an array type
num must be a fixed num (i.e. cannot be dynamic)

For example:
(make-array 5 i64) will return the type |5,i64|*
A pointer to an array holding 5 i64's

Be aware that array types are NOT equivelent to pointer
types.  For example |5,i64| does not equal i64*
In fact for a bitcast you would cast |5,i64|* to i64*
and |5,double|** to double** etc..

array-ref and array-set now work happily with either
pointer types (as before) or a pointer to an array type.

This change also adds heap-alloc and stack-alloc which come
in two varieties:

(heap-alloc type) or (heap-alloc num type)
(stack-alloc type) or (stack-alloc num type)

These both allocate (either on the heap or the stack)
an amount of memory equal to either sizeof(type) or
num * sizeof(type), and return a pointer (type*) to
the newly allocated memory.

heap-alloc and stack-alloc are similar to make-array
however (a) they do not create an array type and (b)
'num' for both heap-alloc and stack-alloc can be
runtime dynamic (unlike make-array).