Windows 7, Boost and LLVM 3 support
This is a major commit (too big sorry!) that touches nearly every file
in the repo.  There are three signficiant changes represented here:

Firstly there are changes to the extempore lang type system to support
upcoming changes in llvm 3.0+.  These are generally behind the scenes
changes but do slightly effect the syntax and semantics of bind-type.

By default extempore still builds (and expects) against the LLVM 2.8
and 2.9 branches.  Also the examples/extempore_lang.scm file is
for the 2.8 and 2.9 branches.  If you want to use LLVM 3.0+ you will
need to compile against a build of the LLVM 3.0+ branch (still in beta)
and set the EXT_LLVM_3 flag:

./all -DEXT_LLVM_3

Additionally there is a patch required to make the LLVM 3.0 branch
compatible.  You'll need to install this patch before you build
LLVM 3.0.  The LLVM src code patch is supplied as a download option
from github: click on downloads on the main extempore github page
and download the llvm 3 patch.

The second major change is to promote cross platform compatibility
by utilizing the boost libraries.  I'm not much of a fan of boost
to be honest but it seems to be the best of a bad bunch of options
at the moment.  This change largely effects networking and threading
code.  As there is quite a lot of threading and networking code
in extempore this change touches on many parts of the system and
potentially adds new bugs/instability.  I hope this is isn't the
case but ...

By default Extempore will still compile with the old code in place
(i.e. pthreads and unix sockets).  If you want to compile against
the boost libraries (instead if pthreads and unix sockets) you
need to set the EXT_BOOST property.

./all -DEXT_BOOST

or if you want boost and LLVM 3

./all -DEXT_BOOST -DEXT_LLVM_3

Finally the boost changes have been made in large part to support
Windows.  Or more specifically 64bit Windows 7 - I'm not going to bother
supporting anything ealier.  It MAY work on other Windows systems,
but I'm not going to be bother testing on them.  Everything currently
works nicely on Windows 7 64bit.  I'm am using this at work at the moment
and so can confirm that it is operational and stable.

There is NO official build support for Windows at the moment.  I will
put a zip file containing my extremetly hacky MSVS 2010 project up
on github (in downloads - same as patch file) for those who want to
build their own.  You will also obviously need to install and build
the Boost, LLVM and PortAudio libraries yourself.  If someone would
like to put together a nice cmake project for all of this that would
be wonderful - until that happens your on your own ;)

Because it's a hastle to build all of this on windows I have also put
a pre-built extempore binary for windows in github downloads.  this
is for convenience only.