Major update part 2
This finishes off some of the changes introduced in the Nov 25 major update.

1) Stack allocation promotion
All stack allocations are now promoted to the head of the function call.  This is to help limit stack usage, particularly for inner loops etc.

2) dotimes semantics changed
dotimes no longer automatically allocates a stack variable. Programmers are now responsible for preallocating variabledefined in the dotimes call.  Additionally, an optional middle argument defines a starting value.  For example:

;; this prints 5-9
(definec test
  (lambda ()
    (let ((i:i64 0))
      (dotimes (i 5 5)
        (printf "%lld\n" i)))))

3) LLVM *must* now be 3.0+.  EXT_LLVM_3 is now longer required.  llvmir3 and llvmti3 have been moved back into llvmir and llvmti respectively.