multichannel input changes (related to 'in' and 'data' of dsp)
'data' (double*) in dsp is now always passed a complete buffer of
audio input samples.  The size of 'data' is bound by the number of
input channels selected by you at runtime and can be dereferenced as
usual using pref.  In other words all input channels are available
through 'data' for each and every output channel. 'in' is still
available but has some specific behaviour as outlined below:

1 channels input:
the 1 channel is passed as 'in' value for every output channel

input and output channels are equal:
each input channel is passed as 'in' to output channel

input and output channels are unqual:
'in' is 0.0 for every output channel.