stack and heap allocation for lambda
Creating dynamic closures has always been
a zone allcoation.  It is now possible to
create a closure on the stack or on the
heap as well as in a zone (which remains
the default 'lambda' behaviour).

lambda now comes in three variants
lambdaz - zone
lambdas - stack
lambdah - heap

lambda = lambdaz (old behaviour)