replace (most) uses of llvm:get-function with impc:ti:closure-exists?
this is tricky in a few parts, because it was checking for _getters - so
if this was being used as some sort of "what stage of the compilation
step am I up to" then things will go wrong, because we're now only
caching whether or not the closure exists (Y/N) in the global list.

E.g. one of the unchanged versions of llvm:get-function is in
impc:ir:compiler (llvmir.xtm:5260), which checks the raw closure name
and then the getter one after the other - not sure whether this will
work after this patch.