Add support for 0-argument lambdas as notes
I find it handy to have a shorthand notes based on the current key/scale, but
that might change over the course of a pattern, and it's also good to be able
to specify accidentals that override the current scale, so simply doing `(qnt
@1)` in the the note function of the looper doesn't work as well as one might
hope.

So... I've tweaked rmap_helper to check if `(eval l)` returns a procedure and,
if so to call that proc. Thinking about this, it should be possible just do
`(if (procedure @1) (@1) @1)` in the player function.

However, one thing that is definitely necessary to make this work is to change
`(eval l)` to `(eval (interaction-environment))`, otherwise it tries to
evaluate the wrong `f`