Extend the pattern language to allow for dotted and shortened notes
A common pattern in traditional tunes is the 𝅘𝅥𝅭 𝅘𝅥𝅮 rhythm, which proved to be
annoyingly hard to express in the ':>' pattern language. You basically have to
replace every note with 'note |', and then do 'dotted-crotchet | | quaver'
which is somewhat annoying.

So, I extended the language to make `note + note -` do the trick (it too me
a while to spot that scheme was never going to let me do `note . note -`.

I've also factored out the code that maps from a note symbol to a number and
which works out where the actual notes are in a pattern and calculates their
positions (both in time and within the pattern list). In theory, this should
allow for the pattern interpretation part of the pattern player to be
parameterised, but I've not really started on that work yet.