Add an xtmtest-with-fixture test and exercise it
Adds an xtmtest-with-fixture macro that sets up a fixture and then
runs its body in the fixture's environment.

There's a bunch of stuff introducing conveniently named macros within
the body, which got added as I was debugging and getting the basic macro
to work and it's a tad tricky to unpick this and present it as a series of
patches adding small pieces of behaviour every time. Sorry.