slight tweak to impc:ti:search-for-dylib
This is to fix a regression on macOS 11 Big Sur, which now no longer has an
OpenGL dylib file in the OpenGL.framework, but instead does some magic stuff on
dlopen() to load the right library. Since the old impc:ti:search-for-dylib
naively checked for file-exists? first (in an effort to print a more helpful
error message) the new "magic" behaviour caused it to break. Hence this patch.